E-19: extract UserTokenResolver from ReflectionMetadataEmitter (#1361)
Extracts the user-type token-resolution band (~1,600 lines) of
ReflectionMetadataEmitter into `internal sealed class UserTokenResolver`.
This is the last big Wave-1 emitter extraction and the consolidation
commit that resolves the temporary couplings E-17/E-18 left pointing into
this band.

Band moved (verbatim bodies, mechanical rewrites only):
- user struct/interface/delegate TypeSpec producers (GetUserStructTypeSpec,
  ResolveUserStructTypeSpecArguments, GetUserInterfaceTypeSpec,
  GetUserDelegateTypeSpec) and their field/method/ctor MemberRef factories
  (GetUserStruct/InterfaceFieldRef, GetUserStruct/InterfaceMethodRef,
  ResolveFieldToken, ResolveInterfaceFieldToken, EncodeOpenMethodSignature,
  EncodeOpenPropertyAccessorSignature, EncodeFunctionReturnSymbol)
- method-token resolvers (ResolveUserInstance/StaticMethodToken,
  ResolveUserInterfaceStatic/InstanceMethodToken,
  ResolveUserPropertyAccessorToken, ResolveClosureInvokeFtnToken,
  ResolveDelegateCtor/InvokeToken, GetStructTypeToken,
  EncodeTypeSymbolIntoSignature, ResolveUserTypeToken)
- ctor-token resolvers (ResolveUserCtorTokenFor{Primary,Default,Explicit},
  ResolveConstructedBase{Explicit,Parameterless}CtorToken)
- generic-user MethodSpec builder + structural type-argument unifier
  (BuildMethodSpecForGeneric{Call,InstanceCall}, BuildMethodSpec,
  InferMethodTypeArgument, TryUnify, TryGetFunctionShapeSlots,
  IsFuncOrActionDefinition, GetGenericTypeArguments, IsValueTupleOpenDefinition,
  StripConversion)
- non-capturing generic-lambda promotion (TryPromoteNonCapturingGenericLambda,
  ResolveLambdaFunctionFtnToken)
- symbolic-substituted-return trio (TryGetSymbolicSubstituted{Property,
  InstanceMethod,ImportedCall}Return) and FunctionTypeNeedsSymbolicDelegate

9 memoization dictionaries moved as private fields of UserTokenResolver
(userStruct{TypeSpec,FieldRef,MethodRef}Cache, userInterface{TypeSpec,
MethodRef,FieldRef}Cache, userDelegate{TypeSpec,CtorRef,InvokeRef}Cache) —
their only users moved with them — plus lambdaMethodTypeArgsByFunction.

Temporary couplings from E-17/E-18 resolved:
- SignatureEncoder (E-17) now calls
  outer.userTokens.ResolveUserStructTypeSpecArguments; the E-17 root widening
  is relocated onto UserTokenResolver.
- ImportedMemberRefFactory (E-18) now calls
  outer.userTokens.GetUserStructTypeSpec / FunctionTypeNeedsSymbolicDelegate.

Kept on the root: the shared static predicates (IsUserGeneric{Type,Interface,
Delegate}Reference, ResolveDelegateTypeArguments, ArgIsSymbolicUserDefined,
TryGetPrimaryCtorTargetField, IsValueTypeSymbol/Nullable) referenced pervasively
as ReflectionMetadataEmitter.X across the collaborators; FindImportedMethod and
EncodeAsyncReturnType (widened private -> internal, reached via the back-ref);
the type-emission helpers EmitDefaultMemberAttributeIfIndexer /
RegisterIndexerAccessorHandles (EmitCore callers). UserTokenResolver is
constructed after signatures + memberRefs (it uses both).

Forwarders left on the root for E-21 (30, called from MethodBodyEmitter
partials / collaborator delegate-injection sites / remaining EmitCore).

Verification: dotnet build GSharp.sln -c Release -graph -> 0 warnings, 0 errors.
PE byte-diff over samples/*.gs: 119/119 identical vs baseline (3 GsharpExtensions*
skipped). RME 8,101 -> 6,213 lines; UserTokenResolver.cs 2,149 lines.
